renamed 'cat sleep' to 'slow cat' and moved logic to constructor for fewer redundant calls
