--- pkg/mapper/examples/cat_sleep/Makefile
@@ -1,22 +1,0 @@
-TAG ?= stable
-PUSH ?= false
-IMAGE_REGISTRY = quay.io/numaio/numaflow-go/map-cat-sleep:${TAG}
-ARCHITECTURES = amd64 arm64
-
-.PHONY: build
-build:
-	for arch in $(ARCHITECTURES); do \
-		CGO_ENABLED=0 GOOS=linux GOARCH=$${arch} go build -v -o ./dist/cat-sleep-example-$${arch} main.go; \
-	done
-
-.PHONY: image-push
-image-push: build
-	docker buildx build -t ${IMAGE_REGISTRY} --platform linux/amd64,linux/arm64 --target cat-sleep . --push
-
-.PHONY: image
-image: build
-	docker build -t ${IMAGE_REGISTRY} --target cat-sleep .
-	@if [ "$(PUSH)" = "true" ]; then docker push ${IMAGE_REGISTRY}; fi
-
-clean:
-	-rm -rf ./dist
