+++ pkg/mapper/examples/slow_cat/Makefile
@@ -1,0 +1,22 @@
+TAG ?= stable
+PUSH ?= false
+IMAGE_REGISTRY = quay.io/numaio/numaflow-go/map-slow-cat:${TAG}
+ARCHITECTURES = amd64 arm64
+
+.PHONY: build
+build:
+	for arch in $(ARCHITECTURES); do \
+		CGO_ENABLED=0 GOOS=linux GOARCH=$${arch} go build -v -o ./dist/slow-cat-example-$${arch} main.go; \
+	done
+
+.PHONY: image-push
+image-push: build
+	docker buildx build -t ${IMAGE_REGISTRY} --platform linux/amd64,linux/arm64 --target slow-cat . --push
+
+.PHONY: image
+image: build
+	docker build -t ${IMAGE_REGISTRY} --target slow-cat .
+	@if [ "$(PUSH)" = "true" ]; then docker push ${IMAGE_REGISTRY}; fi
+
+clean:
+	-rm -rf ./dist
