+++ pkg/mapper/examples/slow_cat/README.md
@@ -1,3 +1,3 @@
-# Cat with Sleep
+# Slow Cat
 
 An example User Defined Function that outputs the same value as the input, but first does a sleep. This example is for the purpose of performing testing with a slow vertex.
