+++ pkg/mapper/examples/slow_cat/go.mod
